Moved statistic class to own directory
