+++ app/views/statistics/_user_table.html.arb
@@ -1,4 +1,4 @@
-collection = statistic.to_table :attributes => [:name, :workgroup, :sign_in_count], :summarize => true
+collection = statistic.to_table :attributes => [:name, :workgroup], :summarize => true
 
 div class: "srutable_wrapper" do
   div id: "user_statistic", class: "srutable_scroll" do
