+++ config/initializers/statistic.rb
@@ -1,0 +1,3 @@
+Dir["#{Rails.root}/lib/statistic/*.rb"].each do |file|
+  require file 
+end
