--- db/migrate/20170316142555_add_owner_index_to_sources.rb
@@ -1,5 +1,0 @@
-class AddOwnerIndexToSources < ActiveRecord::Migration
-  def change
-    add_index :sources, :wf_owner
-  end
-end
